+++ app/Http/Controllers/ResidentArchiveController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\ResidentArchive\UpdateRequest;
 use App\Http\Resources\ResidentArchiveResource;
 use App\Http\Resources\ResidentArchiveResourceCollection;
 use App\Repositories\Contracts\ResidentArchiveRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 
 class ResidentArchiveController extends Controller
@@ -32,9 +33,12 @@ class ResidentArchiveController extends Controller
      *
      * @param IndexRequest $request
      * @return ResidentArchiveResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [ResidentArchive::class, $request->get('propertyId')]);
+
         $residentArchives = $this->residentArchiveRepository->findBy($request->all());
 
         return new ResidentArchiveResourceCollection($residentArchives);
@@ -44,9 +48,13 @@ class ResidentArchiveController extends Controller
      * Store a newly created resource in storage.
      *
      * @param StoreRequest $request
+     * @return ResidentArchiveResourceCollection
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [ResidentArchive::class, $request->get('propertyId')]);
+
         $residentArchives = $this->residentArchiveRepository->saveMultipleResidents($request->all());
 
         return new ResidentArchiveResourceCollection($residentArchives);
@@ -57,9 +65,12 @@ class ResidentArchiveController extends Controller
      *
      * @param ResidentArchive $residentArchive
      * @return ResidentArchiveResource
+     * @throws AuthorizationException
      */
     public function show(ResidentArchive $residentArchive)
     {
+        $this->authorize('show', $residentArchive);
+
         return new ResidentArchiveResource($residentArchive);
     }
 
@@ -69,9 +80,12 @@ class ResidentArchiveController extends Controller
      * @param UpdateRequest $request
      * @param ResidentArchive $residentArchive
      * @return ResidentArchiveResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, ResidentArchive $residentArchive)
     {
+        $this->authorize('update', $residentArchive);
+
         $residentArchive = $this->residentArchiveRepository->update($residentArchive, $request->all());
 
         return new ResidentArchiveResource($residentArchive);
@@ -82,9 +96,12 @@ class ResidentArchiveController extends Controller
      *
      * @param ResidentArchive $residentArchive
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(ResidentArchive $residentArchive)
     {
+        $this->authorize('destroy', $residentArchive);
+
         $this->residentArchiveRepository->delete($residentArchive);
 
         return response()->json(null, 204);
