+++ app/Http/Controllers/ResidentVehicleController.php
@@ -9,6 +9,7 @@ use App\Http\Requests\ResidentVehicle\UpdateRequest;
 use App\Http\Resources\ResidentVehicleResource;
 use App\Http\Resources\ResidentVehicleResourceCollection;
 use App\Repositories\Contracts\ResidentVehicleRepository;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\Request;
 
 class ResidentVehicleController extends Controller
@@ -32,9 +33,12 @@ class ResidentVehicleController extends Controller
      *
      * @param IndexRequest $request
      * @return ResidentVehicleResourceCollection
+     * @throws AuthorizationException
      */
     public function index(IndexRequest $request)
     {
+        $this->authorize('list', [ResidentVehicle::class, $request->get('residentId')]);
+
         $residentVehicles = $this->residentVehicleRepository->findBy($request->all());
 
         return new ResidentVehicleResourceCollection($residentVehicles);
@@ -45,9 +49,12 @@ class ResidentVehicleController extends Controller
      *
      * @param StoreRequest $request
      * @return ResidentVehicleResource
+     * @throws AuthorizationException
      */
     public function store(StoreRequest $request)
     {
+        $this->authorize('store', [ResidentVehicle::class, $request->get('residentId')]);
+
         $residentVehicle = $this->residentVehicleRepository->save($request->all());
 
         return new ResidentVehicleResource($residentVehicle);
@@ -58,9 +65,12 @@ class ResidentVehicleController extends Controller
      *
      * @param ResidentVehicle $residentVehicle
      * @return ResidentVehicleResource
+     * @throws AuthorizationException
      */
     public function show(ResidentVehicle $residentVehicle)
     {
+        $this->authorize('show', $residentVehicle);
+
         return new ResidentVehicleResource($residentVehicle);
     }
 
@@ -70,9 +80,12 @@ class ResidentVehicleController extends Controller
      * @param UpdateRequest $request
      * @param ResidentVehicle $residentVehicle
      * @return ResidentVehicleResource
+     * @throws AuthorizationException
      */
     public function update(UpdateRequest $request, ResidentVehicle $residentVehicle)
     {
+        $this->authorize('update', $residentVehicle);
+
         $residentVehicle = $this->residentVehicleRepository->update($residentVehicle, $request->all());
 
         return new ResidentVehicleResource($residentVehicle);
@@ -83,9 +96,12 @@ class ResidentVehicleController extends Controller
      *
      * @param ResidentVehicle $residentVehicle
      * @return void
+     * @throws AuthorizationException
      */
     public function destroy(ResidentVehicle $residentVehicle)
     {
+        $this->authorize('destroy', $residentVehicle);
+
         $this->residentVehicleRepository->delete($residentVehicle);
 
         return response()->json(null, 204);
